requirements.txt
argparse
requests
beautifulsoup4
python-dotenv
matplotlib
pymongo
colorlog
